% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sindex_AgeSIToHt.R
\docType{methods}
\name{Sindex_AgeSIToHt}
\alias{Sindex_AgeSIToHt}
\alias{Sindex_AgeSIToHt,integer,numeric,integer,numeric,numeric-method}
\alias{Sindex_AgeSIToHt,numeric,numeric,numeric,numeric,numeric-method}
\title{Calculate a height based on an age, site index and site index curve.}
\usage{
Sindex_AgeSIToHt(curve, age, ageType, siteIndex, y2bh)

\S4method{Sindex_AgeSIToHt}{integer,numeric,integer,numeric,numeric}(curve, age,
  ageType, siteIndex, y2bh)

\S4method{Sindex_AgeSIToHt}{numeric,numeric,numeric,numeric,numeric}(curve, age,
  ageType, siteIndex, y2bh)
}
\arguments{
\item{curve}{Integer, The particular site index curve to project the height and age along.}

\item{age}{Numeric, The age of the trees indicated by the curve selection.  The
interpretation of this age is modified by the 'ageType' parameter.}

\item{ageType}{Integer, Age type. Must be one of:
\code{SI_AT_TOTAL}, the age is the total age of the stand in years since
planting, or \code{SI_AT_BREAST}, the age indicates the number of years since the stand
reached breast height.}

\item{siteIndex}{Numeric, The site index value of the stand.}

\item{y2bh}{Numeric, Years to breast height.
The number of years it takes the stand to reach breast height.}
}
\value{
\code{output} the computed height
        \code{error} 0, or an error code under the following conditions:
   
   return value    condition
   ------------    ---------
   SI_ERR_CURVE    input curve is not a valid curve index.
   SI_ERR_GI_MIN   if bhage < 0.5 (for GI)
   SI_ERR_GI_MAX   if bhage > GI range
   SI_ERR_NO_ANS   if computed age > 999
   SI_ERR_GI_TOT   if total age and GI curve
   SI_ERR_LT13     if site index <= 1.3
}
\description{
Converts an Age and Site Index to a Height for a particular Site Index
   Curve.
}
