% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_CurveNotes}
\alias{Sindex_CurveNotes}
\title{Curve notes}
\usage{
Sindex_CurveNotes(cu_index)
}
\arguments{
\item{cu_index}{Integer, Curve index.}
}
\value{
A string containing notes on use of curve.
   If input parameter is not a valid curve index, the return is the
   null pointer.
}
\description{
Returns string containing notes on use.
}
