% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_CurveSource}
\alias{Sindex_CurveSource}
\title{Curve source}
\usage{
Sindex_CurveSource(cu_index)
}
\arguments{
\item{cu_index}{Integer, Curve index.}
}
\value{
String containing publication citation.
   If input parameter is not a valid curve index, the return is the
   null pointer.
}
\description{
Returns string containing publication source.
}
