% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_CurveToSpecies}
\alias{Sindex_CurveToSpecies}
\title{Convert curve index to species index}
\usage{
Sindex_CurveToSpecies(cu_index)
}
\arguments{
\item{cu_index}{Integer, Specifies cuive index.}
}
\value{
Integer species index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition

   ------------    ---------

   SI_ERR_CURVE    input curve is not a valid curve index for any species.
}
\description{
Returns species index for a given curve index.
}
