% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_CurveUse}
\alias{Sindex_CurveUse}
\title{Returns a code telling what functions are available for a curve index.}
\usage{
Sindex_CurveUse(cu_index)
}
\arguments{
\item{cu_index}{Integer, Defines curve index.}
}
\value{
Integer code.
   May return an error code under the following conditions:

   return value    condition

   ------------    ---------

   SI_ERR_CURVE    input curve is not a valid curve index.
}
\description{
Returns a code telling what functions are available for a curve index.
}
\note{
Code bits are set as follows:
     0001: ht = fn (si, age)
     0010: si = fn (ht, age)
     0100: y2bh = fn (si)
     1000: si = fn (ht, age) growth intercept
}
