% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_DefCurveEst}
\alias{Sindex_DefCurveEst}
\title{Returns default curve index for a species and establishment type.}
\usage{
Sindex_DefCurveEst(sp_index, estab)
}
\arguments{
\item{sp_index}{Integer, Speciefies species index.}

\item{estab}{Integer, Speciefies establishment type.}
}
\value{
Integer curve index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
   SI_ERR_ESTAB    input parameter is not a valid establishment type
   SI_ERR_NO_ANS   no curves defined for this species
}
\description{
Returns default curve index for a species and establishment type.
}
