% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_NextCurve}
\alias{Sindex_NextCurve}
\title{Returns next defined curve index for a species.}
\usage{
Sindex_NextCurve(sp_index, cu_index)
}
\arguments{
\item{sp_index}{Integer, Specifies species index.}

\item{cu_index}{Integer, Specifies curve index.}
}
\value{
Integer curve index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition

   ------------    ---------

   SI_ERR_SPEC     input species is not a valid species index.

   SI_ERR_CURVE    input curve is not a valid curve index for this species.

   SI_ERR_NO_ANS   input parameter is last defined index for this species.
}
\description{
Returns next defined curve index for a species.
}
\note{
No assumption should be made about the ordering of the curves.
}
