% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_NextSpecies}
\alias{Sindex_NextSpecies}
\title{Next species defined in Sindex}
\usage{
Sindex_NextSpecies(sp_index)
}
\arguments{
\item{sp_index}{Integer, Specifies species index.}
}
\value{
Integer species index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
   SI_ERR_NO_ANS   input parameter is last defined species index
}
\description{
Given a species index, returns the next species defined in Sindex.
}
\note{
No assumption should be made about the ordering of the species.
}
