% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sindex_SCToSI.R
\docType{methods}
\name{Sindex_SCToSI}
\alias{Sindex_SCToSI}
\alias{Sindex_SCToSI,integer,character,character-method}
\alias{Sindex_SCToSI,numeric,character,character-method}
\title{Get site index based on site class.}
\usage{
Sindex_SCToSI(sp_index, sitecl, fiz)

\S4method{Sindex_SCToSI}{integer,character,character}(sp_index, sitecl, fiz)

\S4method{Sindex_SCToSI}{numeric,character,character}(sp_index, sitecl, fiz)
}
\arguments{
\item{sp_index}{Integer, Species index.}

\item{sitecl}{character, Site class, must be one of \code{G}, \code{M}, \code{P} and \code{L}.}

\item{fiz}{character, Forest inventory zone: (A,B,C)=coast, (D,E,F,G,H,I,J,K,L)=interior.}
}
\value{
\code{output} contains site index;
        \code{error} contains error information, i.e., 
   0, or an error code under the following conditions:
   
   return value    condition
   ------------    ---------
   SI_ERR_SPEC     source species index is not valid, or no conversion
   SI_ERR_CLASS    if site class is unknown
   SI_ERR_FIZ      if FIZ code is unknown
}
\description{
Get site index based on site class.
}
