% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_SIToSI}
\alias{Sindex_SIToSI}
\title{Site index conversion between species}
\usage{
Sindex_SIToSI(sp_index1, site, sp_index2)
}
\arguments{
\item{sp_index1}{Integer, Source species index.}

\item{site}{Numeric, Source species site index.}

\item{sp_index2}{Integer, Target species index.}
}
\value{
Floating point target species site index. (computed)

 Return Value
 ------------
   0, or an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     source or target species index is not valid.
   SI_ERR_NO_ANS   there is no conversion defined.
}
\description{
Site index conversion between species. This function needs to be revised.
}
