% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesMapRemap.R
\name{Sindex_SpecMap}
\alias{Sindex_SpecMap}
\title{Determine species index from species code}
\usage{
Sindex_SpecMap(sc)
}
\arguments{
\item{sc}{Character, Defined as species code.}
}
\value{
Species index.
   May return an error code under the following conditions:
   
   return value    condition
   ------------    ---------
   SI_ERR_CODE     if species code is unknown
}
\description{
Determine species index from species code
}
\note{
Species code string can be 1, 2, or 3 letters; upper/lower case
     is ignored.
}
