% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesMapRemap.R
\name{Sindex_SpecRemap}
\alias{Sindex_SpecRemap}
\title{Remap species to recommended species, and return species index}
\usage{
Sindex_SpecRemap(sc, fiz)
}
\arguments{
\item{sc}{Character, Species code.}

\item{fiz}{Character. Forest inventory zone: (A,B,C)=coast,
                      (D,E,F,G,H,I,J,K,L)=interior.

 @return Species index.
   May return an error code under the following conditions:
   
   return value    condition
   ------------    ---------
   SI_ERR_CODE     if species code is unknown
   SI_ERR_FIZ      if FIZ code is unknown}
}
\description{
Remap species to recommended species, and return species index
}
\note{
Species code string can be 1, 2, or 3 letters; upper/lower case
     is ignored.  FIZ is only used where needed, such as for species
     code "FD".
}
