% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sindex_SpecUse}
\alias{Sindex_SpecUse}
\title{Returns a code telling where a species generally exists.}
\usage{
Sindex_SpecUse(sp_index)
}
\arguments{
\item{sp_index}{Integer, Specifies species index.}
}
\value{
Integer code.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
}
\description{
Returns a code telling where a species generally exists.
}
\note{
Code bits are set as follows:
        1: BC coast
       10: BC interior
      100: common species in BC (0 means uncommon)
}
