% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sindex_Y2BH.R
\docType{methods}
\name{Sindex_Y2BH}
\alias{Sindex_Y2BH}
\alias{Sindex_Y2BH,integer,numeric-method}
\alias{Sindex_Y2BH,numeric,numeric-method}
\title{Calculates the number of years a stand takes to grow from seed to
    breast height.}
\usage{
Sindex_Y2BH(curve, siteIndex)

\S4method{Sindex_Y2BH}{integer,numeric}(curve, siteIndex)

\S4method{Sindex_Y2BH}{numeric,numeric}(curve, siteIndex)
}
\arguments{
\item{curve, }{Integer, The particular site index curve to project the height and age along.}

\item{siteIndex, }{Numeric, The site index value of the stand.}
}
\value{
\code{output}: computed years to breast height.
        \code{error}: error information, i.e.,
            0, or an error code under the following conditions:
           return value    condition
   ------------    ---------
   SI_ERR_CURVE    input curve is not a valid curve index.
   SI_ERR_GI_TOT   if GI curve
   SI_ERR_LT13     if site index <= 1.3
}
\description{
Calculates the number of years a stand takes to grow from seed to
    breast height.
}
