% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_AgeToAge.R
\name{SIndexR_AgeToAge}
\alias{SIndexR_AgeToAge}
\title{Convert age to other type}
\usage{
SIndexR_AgeToAge(cu_index, age1, age_type1, y2bh, age_type2)
}
\arguments{
\item{cu_index}{Integer, Curve index.}

\item{age1}{Numeric, Source age.}

\item{age_type1}{Integer, Type of source age (SI_AT_BREAST or SI_AT_TOTAL).}

\item{y2bh}{Numeric, Years to breast height.}

\item{age_type2}{Integer, Type of target age (SI_AT_BREAST or SI_AT_TOTAL).}
}
\value{
\code{output} contains computed targe age.
 \code{error} contains error information, i.e.,
     0, or an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_CURVE    input curve is not a valid curve index for this species.
   SI_ERR_AGE_TYPE unknown age type.
}
\description{
Age conversion between age types (total vs breast height)
}
