% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_CurveName.R
\name{SIndexR_CurveName}
\alias{SIndexR_CurveName}
\title{Curve name}
\usage{
SIndexR_CurveName(cu_index)
}
\arguments{
\item{cu_index}{Integer/Numeric, Curve index.}
}
\value{
A string containing curve author and date.
   If input parameter is not a valid curve index, the return is the
   \code{null}.
}
\description{
Returns string containing author and date of curve.
}
\note{
Curve name string examples: "Bruce (1981)", "Nigh (1998)".
}
