% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_CurveSource.R
\name{SIndexR_CurveSource}
\alias{SIndexR_CurveSource}
\title{Curve source}
\usage{
SIndexR_CurveSource(cu_index)
}
\arguments{
\item{cu_index}{Integer/Numeric, Curve index.}
}
\value{
String containing publication citation.
   If input parameter is not a valid curve index, the return is the
   null pointer.
}
\description{
Returns string containing publication source.
}
