% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_DefCurve.R
\name{SIndexR_DefCurve}
\alias{SIndexR_DefCurve}
\title{Returns default curve index for a species}
\usage{
SIndexR_DefCurve(sp_index)
}
\arguments{
\item{sp_index}{Integer, Specifies species index.}
}
\value{
Integer curve index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
   SI_ERR_NO_ANS   no curves defined for this species
}
\description{
Returns default curve index for a species
}
