% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_DefGICurve.R
\name{SIndexR_DefGICurve}
\alias{SIndexR_DefGICurve}
\title{Returns default GI curve index for a species.}
\usage{
SIndexR_DefGICurve(sp_index)
}
\arguments{
\item{sp_index}{Integer/Numeric, Specifies species index.}
}
\value{
Integer curve index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
   SI_ERR_NO_ANS   no GI equations defined for this species.
}
\description{
Returns default GI curve index for a species.
}
