% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_FirstCurve.R
\name{SIndexR_FirstCurve}
\alias{SIndexR_FirstCurve}
\title{Returns first defined curve index for a species.}
\usage{
SIndexR_FirstCurve(sp_index)
}
\arguments{
\item{sp_index}{Integer, Specifies species index.}
}
\value{
Integer curve index, for use in other Sindex functions.
   May return an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_SPEC     input parameter is not a valid species index
   SI_ERR_NO_ANS   there are no curves defined for this species
}
\description{
Returns first defined curve index for a species.
}
\note{
No assumption should be made about the ordering of the curves.
}
