% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sindex_HtAgeToSI.R
\name{SIndexR_HtAgeToSI}
\alias{SIndexR_HtAgeToSI}
\title{Calculates site index based on height, age and site index curve.}
\usage{
SIndexR_HtAgeToSI(curve, age, ageType, height, estType)
}
\arguments{
\item{curve}{Integer/Numeric, Specifies site index curve.
The particular site index curve to project the height and age along.}

\item{age}{numeric, Tree age.The age of the trees indicated by the curve selection.  The
interpretation of this age is modified by the 'ageType' parameter.}

\item{ageType}{Integer/Numeric, Defines age type. Must be one of:
\code{SI_AT_TOTAL}, the age is the total age of the stand in years since
planting; \code{SI_AT_BREAST}, the age indicates the number of years since the stand
reached breast height.}

\item{height}{numeric, The height of the species in metres.}

\item{estType}{Integer/Numeric, Defines estimate type. Must be one of:
v             \code{SI_EST_DIRECT}, compute the site index based on direct equations
              if available.  If the equations are not available,
              then automatically fall to the \code{SI_EST_ITERATE}
              method; \code{SI_EST_ITERATE}, compute the site index based on an iterative
              method which converges on the true site index.}
}
\value{
\code{output} contains computed site index.
        \code{error} contains error values.

     If an error condition occurs, the site index is set to the
       same as the returned site index value.

 Return Value
 ------------
   0, or an error code under the following conditions:

   return value    condition
   ------------    ---------
   SI_ERR_CURVE    input curve is not a valid curve index.
   SI_ERR_GI_MIN   if bhage < 0.5
   SI_ERR_GI_MAX   if bhage > GI range
   SI_ERR_NO_ANS   if computed SI > 999
   SI_ERR_GI_TOT   if total age and GI curve
}
\description{
Converts a Height and Age to a Site Index for a particular Site Index
             Curve.
}
