% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_SpecName.R
\name{SIndexR_SpecName}
\alias{SIndexR_SpecName}
\title{Returns string containing species name.}
\usage{
SIndexR_SpecName(sp_index)
}
\arguments{
\item{sp_index}{Integer/Numeric, Specifies species index}
}
\value{
Pointer to string containing species name.
   If input parameter is not a valid species index, the return is the
   null pointer.
}
\description{
Returns string containing species name.
}
\note{
Species name string examples: "Coastal Douglas-fir", "Sitka Spruce".
}
