% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIndexR_FirstSpecies.R
\name{SIndexR_FirstSpecies}
\alias{SIndexR_FirstSpecies}
\title{First species defined in Sindex}
\usage{
SIndexR_FirstSpecies()
}
\value{
Integer species index, for use in other Sindex functions.
}
\description{
Returns a species index for the first species defined in Sindex
}
\note{
No assumption should be made about the ordering of the species.
}
